refactor(user): improve validation during User object creation
This commit enhances field validation when creating a User object.

Additionally, it removes the manual assignment of a UUID within the `initCommonFields` method. This change prevents Hibernate from mistakenly attempting an UPDATE instead of an INSERT when persisting a new user, as it previously detected the entity as already having an ID. By ensuring the ID field is NULL for new User objects passed to the persistence layer, JPA can correctly generate the UUID upon insertion.